==> 02_Data_Masking.ipynb <==
import snowflake.snowpark as snp
from snowflake.snowpark import functions as F

## Data Masking
We don't want all the users to see the age of our squirrels. We want only the internal devs to access that sensitive data.

Therefore we are going to create a masking policy to dynamically protect this PII column

### 1. Load credentials and connect to Snowflake

from steps.snowpark_connection import snowpark_connect
session, state_dict = snowpark_connect('./include/state.json')
session.use_warehouse(state_dict['compute_parameters']['default_warehouse'])

session.table(state_dict['table_name']).count()

### 2. Create masking policy to protect our PII data

try:
  session.sql("CREATE OR REPLACE MASKING POLICY mask_string_age AS \
      (val STRING) RETURNS STRING -> \
      CASE \
      WHEN CURRENT_ROLE() in ('ACCOUNTADMIN') THEN val \
        ELSE '**masked**'\
      END;").collect()
except:
    pass

Apply the masking policy to the column

session.sql("ALTER TABLE "+ state_dict['table_name'] +" MODIFY COLUMN AGE SET MASKING POLICY mask_string_age;").collect()

session.sql("grant usage on database "+ state_dict['connection_parameters']['database'] +" to role PUBLIC;").collect()
session.sql("grant usage on schema " + state_dict['connection_parameters']['database'] + "."+ state_dict['connection_parameters']['schema']+ " to role PUBLIC;").collect()
session.sql("grant select on table "+ state_dict['connection_parameters']['database'] + "."+ state_dict['connection_parameters']['schema']+ "." +  state_dict['table_name'] + " to role PUBLIC;").collect()


See the masking policy in action. And a bit of Snowpark magic

session.use_role(state_dict['additional_roles']['default_role'])
session.table(state_dict['table_name']).select(F.col("age")).limit(5).show()

session.use_role(state_dict['additional_roles']['project_role'])
session.table(state_dict['table_name']).select(F.col("age")).limit(5).show()

### 3. (Optional) Behind the scenes

# (optional) behind the scenes
session.sql("SELECT query_type,query_text "+
            "FROM TABLE(information_schema.query_history()) "+
            "ORDER BY start_time DESC "+
            "LIMIT 2").show(max_width=200)

session.close()
