==> 03_Data_Exploration.ipynb <==
import snowflake.snowpark as snp
from snowflake.snowpark import functions as F
from snowflake.snowpark import types as T
import pandas as pd
import seaborn as sns
import matplotlib.pyplot as plt

# Data Exploration
In this example notebook we show how to start data exploration using popular python libraries, such as pandas or seaborn. 

### 1. Load the Credentials

from steps.snowpark_connection import snowpark_connect
session, state_dict = snowpark_connect('./include/state.json')

### 2. Load Data

snowdf = session.table(state_dict['table_name'])
print("Count records: " + str(snowdf.count()))
snowdf.columns

### 3. Sample Data and explore

df = snowdf.sample(n=500).to_pandas()

df.head()

Below are just some basic plots

plt.style.use('seaborn-darkgrid')
plt.figure(figsize=(6,3))
df.groupby(["AGE"]).size().plot(kind='bar');


plt.figure(figsize=(4,2))
activity_data = (
    df[["RUNNING", "CHASING", "CLIMBING", "EATING", "FORAGING"]].sum().reset_index()
    )
activity_data.columns = ["Activity", "Count"]
activity_data.groupby(["Activity"]).agg({'Count':'sum'}).plot(kind='barh');

plt.figure(figsize=(6,3))
df.groupby(["PRIMARY_FUR_COLOR"]).size().plot(kind='area');

session.close()
